+++ src/main/java/com/hyun/udong/udong/application/service/UdongService.java
@@ -9,10 +9,10 @@ import com.hyun.udong.travelschedule.infrastructure.repository.CityRepository;
 import com.hyun.udong.udong.domain.*;
 import com.hyun.udong.udong.infrastructure.repository.ParticipantRepository;
 import com.hyun.udong.udong.infrastructure.repository.UdongRepository;
-import com.hyun.udong.udong.presentation.dto.CreateUdongRequest;
-import com.hyun.udong.udong.presentation.dto.CreateUdongResponse;
-import com.hyun.udong.udong.presentation.dto.FindUdongsCondition;
-import com.hyun.udong.udong.presentation.dto.SimpleUdongResponse;
+import com.hyun.udong.udong.presentation.dto.request.CreateUdongRequest;
+import com.hyun.udong.udong.presentation.dto.request.FindUdongsCondition;
+import com.hyun.udong.udong.presentation.dto.response.CreateUdongResponse;
+import com.hyun.udong.udong.presentation.dto.response.SimpleUdongResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
