+++ src/main/java/com/hyun/udong/udong/infrastructure/repository/UdongRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.hyun.udong.udong.infrastructure.repository;
 
 import com.hyun.udong.udong.domain.Udong;
-import com.hyun.udong.udong.presentation.dto.FindUdongsCondition;
+import com.hyun.udong.udong.presentation.dto.request.FindUdongsCondition;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
