+++ src/main/java/com/hyun/udong/udong/presentation/controller/UdongController.java
@@ -4,10 +4,10 @@ import com.hyun.udong.common.annotation.LoginMember;
 import com.hyun.udong.common.dto.PagedResponse;
 import com.hyun.udong.member.domain.Member;
 import com.hyun.udong.udong.application.service.UdongService;
-import com.hyun.udong.udong.presentation.dto.CreateUdongRequest;
-import com.hyun.udong.udong.presentation.dto.CreateUdongResponse;
-import com.hyun.udong.udong.presentation.dto.FindUdongsCondition;
-import com.hyun.udong.udong.presentation.dto.SimpleUdongResponse;
+import com.hyun.udong.udong.presentation.dto.request.CreateUdongRequest;
+import com.hyun.udong.udong.presentation.dto.request.FindUdongsCondition;
+import com.hyun.udong.udong.presentation.dto.response.CreateUdongResponse;
+import com.hyun.udong.udong.presentation.dto.response.SimpleUdongResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
