+++ src/main/java/com/hyun/udong/udong/presentation/dto/response/ParticipantResponse.java
@@ -1,4 +1,4 @@
-package com.hyun.udong.udong.presentation.dto;
+package com.hyun.udong.udong.presentation.dto.response;
 
 import com.hyun.udong.member.domain.Member;
