+++ src/test/java/com/hyun/udong/udong/application/service/UdongServiceTest.java
@@ -8,10 +8,10 @@ import com.hyun.udong.member.domain.SocialType;
 import com.hyun.udong.member.infrastructure.repository.MemberRepository;
 import com.hyun.udong.udong.domain.Udong;
 import com.hyun.udong.udong.infrastructure.repository.UdongRepository;
-import com.hyun.udong.udong.presentation.dto.CreateUdongRequest;
-import com.hyun.udong.udong.presentation.dto.CreateUdongResponse;
-import com.hyun.udong.udong.presentation.dto.FindUdongsCondition;
-import com.hyun.udong.udong.presentation.dto.SimpleUdongResponse;
+import com.hyun.udong.udong.presentation.dto.request.CreateUdongRequest;
+import com.hyun.udong.udong.presentation.dto.request.FindUdongsCondition;
+import com.hyun.udong.udong.presentation.dto.response.CreateUdongResponse;
+import com.hyun.udong.udong.presentation.dto.response.SimpleUdongResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
